move drag event handlers to dnd-placeholder
